product service converted soap to rest
