--- catalog-service/src/main/java/com/cloudtemplate/catalogservice/CatalogServiceApplication.java
@@ -1,75 +1,0 @@
-package com.cloudtemplate.catalogservice;
-
-import com.cloudtemplate.shared.util.DefaultProfileUtil;
-import org.apache.commons.lang.StringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.context.annotation.Bean;
-import org.springframework.core.env.Environment;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.net.InetAddress;
-import java.net.UnknownHostException;
-import java.util.concurrent.Executor;
-
-@SpringBootApplication
-@EnableAsync
-@EnableDiscoveryClient
-@EnableCaching
-public class CatalogServiceApplication {
-
-    private static final Logger log = LoggerFactory.getLogger(CatalogServiceApplication.class);
-
-    public static void main(String[] args) {
-        SpringApplication app = new SpringApplication(CatalogServiceApplication.class);
-        DefaultProfileUtil.addDefaultProfile(app);
-        Environment env = app.run(args).getEnvironment();
-        logApplicationStartup(env);
-    }
-
-    private static void logApplicationStartup(Environment env) {
-        String protocol = "http";
-        if (env.getProperty("server.ssl.key-store") != null) {
-            protocol = "https";
-        }
-        String serverPort = env.getProperty("server.port");
-        String contextPath = env.getProperty("server.servlet.context-path");
-        if (StringUtils.isBlank(contextPath)) {
-            contextPath = "/";
-        }
-        String hostAddress = "localhost";
-        try {
-            hostAddress = InetAddress.getLocalHost().getHostAddress();
-        } catch (UnknownHostException e) {
-            log.error("The host name could not be determined, using `localhost` as fallback");
-        }
-        log.info("\n----------------------------------------------------------\n\t" +
-                        "Application '{}' is running! Access URLs:\n\t" +
-                        "Local: \t\t{}://localhost:{}{}\n\t" +
-                        "External: \t{}://{}:{}{}\n\t" + "\n----------------------------------------------------------",
-                env.getProperty("spring.application.name"),
-                protocol,
-                serverPort,
-                contextPath,
-                protocol,
-                hostAddress,
-                serverPort,
-                contextPath);
-    }
-
-    @Bean
-    public Executor asyncExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(5);
-        executor.setMaxPoolSize(5);
-        executor.setQueueCapacity(500);
-        executor.setThreadNamePrefix("Asynchronous Process-");
-        executor.initialize();
-        return executor;
-    }
-}
