--- catalog-service/src/main/java/com/cloudtemplate/catalogservice/bus/LogProducerManager.java
@@ -1,7 +1,0 @@
-package com.cloudtemplate.catalogservice.bus;
-
-public interface LogProducerManager {
-    void info(String methodName, Object request, Object respone);
-
-    void error(String methodName, Object request, Object respone);
-}
