--- catalog-service/src/main/java/com/cloudtemplate/catalogservice/config/LoggingAspectConfiguration.java
@@ -1,17 +1,0 @@
-package com.cloudtemplate.catalogservice.config;
-
-import com.cloudtemplate.catalogservice.aop.logging.LoggingAspect;
-import com.cloudtemplate.catalogservice.bus.LogProducerManager;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-
-@Configuration
-@EnableAspectJAutoProxy
-public class LoggingAspectConfiguration {
-
-    @Bean
-    public LoggingAspect loggingAspect(LogProducerManager logProducerManager) {
-        return new LoggingAspect(logProducerManager);
-    }
-}
