--- catalog-service/src/main/java/com/cloudtemplate/catalogservice/config/WebServiceConfig.java
@@ -1,41 +1,0 @@
-package com.cloudtemplate.catalogservice.config;
-
-import org.springframework.boot.web.servlet.ServletRegistrationBean;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.ws.config.annotation.EnableWs;
-import org.springframework.ws.transport.http.MessageDispatcherServlet;
-import org.springframework.ws.wsdl.wsdl11.DefaultWsdl11Definition;
-import org.springframework.xml.xsd.SimpleXsdSchema;
-import org.springframework.xml.xsd.XsdSchema;
-
-import static com.cloudtemplate.shared.constans.ApplicationConstants.CUSTOMER_ENDPOINT;
-
-@EnableWs
-@Configuration
-public class WebServiceConfig {
-    @Bean
-    public ServletRegistrationBean<MessageDispatcherServlet> messageDispatcherServlet(ApplicationContext context) {
-        MessageDispatcherServlet messageDispatcherServlet = new MessageDispatcherServlet();
-        messageDispatcherServlet.setApplicationContext(context);
-        messageDispatcherServlet.setTransformWsdlLocations(true);
-        return new ServletRegistrationBean<>(messageDispatcherServlet, "/ws/*");
-    }
-
-    @Bean(name = "catalogs")
-    public DefaultWsdl11Definition defaultWsdl11Definition(XsdSchema customerSchema) {
-        DefaultWsdl11Definition definition = new DefaultWsdl11Definition();
-        definition.setPortTypeName("CatalogPort");
-        definition.setTargetNamespace(CUSTOMER_ENDPOINT);
-        definition.setLocationUri("/ws");
-        definition.setSchema(customerSchema);
-        return definition;
-    }
-
-    @Bean
-    public XsdSchema customerSchema() {
-        return new SimpleXsdSchema(new ClassPathResource("xsd/product.xsd"));
-    }
-}
