--- catalog-service/src/main/java/com/cloudtemplate/catalogservice/dto/CustomAllProductResponse.java
@@ -1,12 +1,0 @@
-package com.cloudtemplate.catalogservice.dto;
-
-import com.crmpoc.catalog.GetAllProductResponse;
-import com.crmpoc.catalog.ProductDetails;
-
-import java.util.List;
-
-public class CustomAllProductResponse extends GetAllProductResponse {
-    public void setProducts(List<ProductDetails> productDetails) {
-        this.productDetails = productDetails;
-    }
-}
