--- catalog-service/src/main/java/com/cloudtemplate/catalogservice/repository/ProductRepository.java
@@ -1,7 +1,0 @@
-package com.cloudtemplate.catalogservice.repository;
-
-import com.cloudtemplate.catalogservice.domain.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
