--- catalog-service/src/main/java/com/crmpoc/catalog/ProductDetails.java
@@ -1,173 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.7 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.01.23 at 01:17:02 PM EET 
-//
-
-
-package com.crmpoc.catalog;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlType;
-
-
-/**
- * <p>Java class for ProductDetails complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>
- * &lt;complexType name="ProductDetails">
- *   &lt;complexContent>
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       &lt;sequence>
- *         &lt;element name="id" type="{http://www.w3.org/2001/XMLSchema}long"/>
- *         &lt;element name="name" type="{http://www.w3.org/2001/XMLSchema}string"/>
- *         &lt;element name="price" type="{http://www.w3.org/2001/XMLSchema}float"/>
- *         &lt;element name="currency" type="{http://www.w3.org/2001/XMLSchema}string"/>
- *         &lt;element name="stock" type="{http://www.w3.org/2001/XMLSchema}long"/>
- *         &lt;element name="enable" type="{http://www.w3.org/2001/XMLSchema}boolean"/>
- *       &lt;/sequence>
- *     &lt;/restriction>
- *   &lt;/complexContent>
- * &lt;/complexType>
- * </pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "ProductDetails", propOrder = {
-    "id",
-    "name",
-    "price",
-    "currency",
-    "stock",
-    "enable"
-})
-public class ProductDetails {
-
-    protected long id;
-    @XmlElement(required = true)
-    protected String name;
-    protected float price;
-    @XmlElement(required = true)
-    protected String currency;
-    protected long stock;
-    protected boolean enable;
-
-    /**
-     * Gets the value of the id property.
-     * 
-     */
-    public long getId() {
-        return id;
-    }
-
-    /**
-     * Sets the value of the id property.
-     * 
-     */
-    public void setId(long value) {
-        this.id = value;
-    }
-
-    /**
-     * Gets the value of the name property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getName() {
-        return name;
-    }
-
-    /**
-     * Sets the value of the name property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setName(String value) {
-        this.name = value;
-    }
-
-    /**
-     * Gets the value of the price property.
-     * 
-     */
-    public float getPrice() {
-        return price;
-    }
-
-    /**
-     * Sets the value of the price property.
-     * 
-     */
-    public void setPrice(float value) {
-        this.price = value;
-    }
-
-    /**
-     * Gets the value of the currency property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getCurrency() {
-        return currency;
-    }
-
-    /**
-     * Sets the value of the currency property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setCurrency(String value) {
-        this.currency = value;
-    }
-
-    /**
-     * Gets the value of the stock property.
-     * 
-     */
-    public long getStock() {
-        return stock;
-    }
-
-    /**
-     * Sets the value of the stock property.
-     * 
-     */
-    public void setStock(long value) {
-        this.stock = value;
-    }
-
-    /**
-     * Gets the value of the enable property.
-     * 
-     */
-    public boolean isEnable() {
-        return enable;
-    }
-
-    /**
-     * Sets the value of the enable property.
-     * 
-     */
-    public void setEnable(boolean value) {
-        this.enable = value;
-    }
-
-}
