--- catalog-service/src/main/java/com/crmpoc/catalog/SaveProductRequest.java
@@ -1,71 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.7 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.01.23 at 01:17:02 PM EET 
-//
-
-
-package com.crmpoc.catalog;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.XmlType;
-
-
-/**
- * <p>Java class for anonymous complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>
- * &lt;complexType>
- *   &lt;complexContent>
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       &lt;sequence>
- *         &lt;element name="ProductDetails" type="{http://www.crmpoc.com/catalog}ProductDetails"/>
- *       &lt;/sequence>
- *     &lt;/restriction>
- *   &lt;/complexContent>
- * &lt;/complexType>
- * </pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "", propOrder = {
-    "productDetails"
-})
-@XmlRootElement(name = "SaveProductRequest")
-public class SaveProductRequest {
-
-    @XmlElement(name = "ProductDetails", required = true)
-    protected ProductDetails productDetails;
-
-    /**
-     * Gets the value of the productDetails property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link ProductDetails }
-     *     
-     */
-    public ProductDetails getProductDetails() {
-        return productDetails;
-    }
-
-    /**
-     * Sets the value of the productDetails property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link ProductDetails }
-     *     
-     */
-    public void setProductDetails(ProductDetails value) {
-        this.productDetails = value;
-    }
-
-}
