--- catalog-service/src/main/java/com/crmpoc/catalog/package-info.java
@@ -1,9 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.7 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.01.23 at 01:17:02 PM EET 
-//
-
-@javax.xml.bind.annotation.XmlSchema(namespace = "http://www.crmpoc.com/catalog", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
-package com.crmpoc.catalog;
