--- catalog-service/src/main/resources/application.yml
@@ -1,59 +1,0 @@
-spring:
-  application:
-    name: catalog-service
-  datasource:
-    hikari:
-      connection-timeout: 60000
-      maximum-pool-size: 5
-    dbcp2:
-      test-while-idle: true
-      validationQuery: SELECT 1
-      timeBetweenEvictionRunsMillis: 60000
-    initialization-mode: always
-    tomcat:
-      max-wait: 15000
-      max-active: 50
-      test-on-borrow: true
-  jpa:
-    properties:
-      connection.characterEncoding: utf-8
-      connection.CharSet: utf-8
-      connection.useUnicode: true
-      hibernate:
-        dialect: org.hibernate.dialect.MySQL5Dialect
-    show-sql: true
-    hibernate:
-      ddl-auto: update
-  kafka:
-    producer:
-      key-serializer: org.apache.kafka.common.serialization.StringSerializer
-      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
-
-server:
-  port: 9080
-  servlet:
-    session:
-      cookie:
-        http-only: true
-
-eureka:
-  instance:
-    instance-id: ${spring.application.name}-${random.uuid}
-  client:
-    service-url:
-      defaultZone: http://${eureka.server.host}:${eureka.server.port}/eureka/
-    register-with-eureka: true
-    fetch-registry: true
-  server:
-    host: localhost
-    port: 8090
-
-kafka:
-  topic:
-    log: system-log
-
-management:
-  endpoint:
-    health:
-      enabled: true
-      show-details: always
