--- catalog-service/src/test/java/com/cloudtemplate/catalogservice/CatalogServiceApplicationTests.java
@@ -1,14 +1,0 @@
-package com.cloudtemplate.catalogservice;
-
-import org.junit.jupiter.api.Test;
-import org.junit.runner.RunWith;
-import org.mockito.junit.MockitoJUnitRunner;
-
-@RunWith(MockitoJUnitRunner.class)
-class CatalogServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
