+++ product-service/entrypoint.sh
