+++ product-service/src/main/fabric8/deployment.yml
