+++ product-service/src/main/fabric8/service.yml
