+++ product-service/src/main/java/com/cloudtemplate/productservice/bus/LogProducerManager.java
@@ -1,0 +1,7 @@
+package com.cloudtemplate.productservice.bus;
+
+public interface LogProducerManager {
+    void info(String methodName, Object request, Object respone);
+
+    void error(String methodName, Object request, Object respone);
+}
