+++ product-service/src/main/java/com/cloudtemplate/productservice/config/LoggingAspectConfiguration.java
@@ -1,0 +1,17 @@
+package com.cloudtemplate.productservice.config;
+
+import com.cloudtemplate.productservice.aop.logging.LoggingAspect;
+import com.cloudtemplate.productservice.bus.LogProducerManager;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.EnableAspectJAutoProxy;
+
+@Configuration
+@EnableAspectJAutoProxy
+public class LoggingAspectConfiguration {
+
+    @Bean
+    public LoggingAspect loggingAspect(LogProducerManager logProducerManager) {
+        return new LoggingAspect(logProducerManager);
+    }
+}
