+++ product-service/src/main/resources/application-docker.yml
