+++ product-service/src/test/java/com/cloudtemplate/productservice/CatalogServiceApplicationTests.java
@@ -1,0 +1,14 @@
+package com.cloudtemplate.productservice;
+
+import org.junit.jupiter.api.Test;
+import org.junit.runner.RunWith;
+import org.mockito.junit.MockitoJUnitRunner;
+
+@RunWith(MockitoJUnitRunner.class)
+class CatalogServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
