+++ shared/src/main/java/com/cloudtemplate/shared/dto/product/OrderFulfilmentRequest.java
@@ -1,0 +1,8 @@
+package com.cloudtemplate.shared.dto.product;
+
+import lombok.Data;
+
+@Data
+public class OrderFulfilmentRequest {
+    private Long productId;
+}
